remove the old pattern explorer that used local patterns
