+++ packages/block-editor/src/components/inserter/block-patterns-explorer/use-patterns-from-directory.js
