+++ packages/block-editor/src/components/inserter/block-patterns-tab.js
@@ -27,7 +27,6 @@ import usePatternsState from './hooks/use-patterns-state';
 import BlockPatternList from '../block-patterns-list';
 import PatternsExplorerModal from './block-patterns-explorer/explorer';
 import MobileTabNavigation from './mobile-tab-navigation';
-import PatternsExplorerModal2 from './block-patterns-explorer2/explorer';
 
 function usePatternsCategories( rootClientId ) {
 	const [ allPatterns, allCategories ] = usePatternsState(
@@ -260,13 +259,6 @@ function BlockPatternsTabs( {
 					initialCategory={ selectedCategory }
 					patternCategories={ categories }
 					onModalClose={ () => setShowPatternsExplorer( false ) }
-				/>
-			) }
-			{ showPatternsExplorer2 && (
-				<PatternsExplorerModal2
-					initialCategory={ selectedCategory }
-					patternCategories={ categories }
-					onModalClose={ () => setShowPatternsExplorer2( false ) }
 				/>
 			) }
 		</>
